fix#1

+++ Helpers.cpp
@@ -1,7 +1,8 @@
 #include "Helpers.h"
 #include <math.h>
 
-double powSum(int a, int b) {
+double powSum(int a, int b) 
+{
 	double exponent = 2;
-	return pow(a + b, exponent);
+	return pow(static_cast<double>(a + b), exponent);
 }
